Improve utc offset heuristic and test it
This improves the UTC offset heuristic. Since the TZif data doesn't
split out what part of the time offset gets you to standard time for a
location and what part accounts for DST, this is figured out using a set
of heuristics.

Previously, all UTC offset discrepancies with Tz were ignored. I.e., if
the aggregate offset was correct, then pass. Now UTC offset
discrepancies are compared and if there's a difference, the test fails.
Some discrepancies are hard to fix so these are allowed. This includes
Monoco and Paris is the 1940s and Morocco. Notes were added describing
the reasons for the discrepancies.