Bump test tzdb to 2021a
The tests pass here. This is intended to provide a more recent check
point that timezone offsets get calculated properly.

Note: I would have updated to 2021c, but tz didn't ship with that
version so the exhaustive check that the regression tests would have
picked up expected differences. 2021a is pretty close, though. 2021d is
the location of the `zic` update. See #19.