Keep tzdb version to avoid confusion
When building zic for testing, this keeps the tzdb version as part of
the directory. The benefit is that when you change the version, you can
have some confidence that you're using the right version of zic.
Previously zic wouldn't be rebuilt and the easiest workaround was to rm
the _build directory.